+++ RELEASE_NOTES.md
@@ -11,6 +11,7 @@ More details at <a href="https://nighttime-imaging.eu/donate/" target="_blank">n
 - Fixed QHY Camera selection being lost on rescan
 - Imaging Platesolve Panel will preset initial gain from platesolve settings
 - Flat Wizard not showing results until N.I.N.A. restarted for some filters
+- Flat Wizard Sky Flat Mode no longer closes a flat device cover when connected
 
 ## Improvements
 - Guide Chart Colors can now be customized in the guider tab